genflow/requirements/base.txt
# SHA1:fdbb0ac89742bb972c23f9d05d19f291afc795a0
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
asgiref==3.8.1
    # via
    #   channels
    #   daphne
    #   django
    #   django-cors-headers
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via daphne
automat==24.8.1
    # via twisted
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
channels[daphne]==4.2.0
    # via -r genflow/requirements/base.in
charset-normalizer==3.4.1
    # via requests
constantly==23.10.4
    # via twisted
cryptography==44.0.2
    # via
    #   autobahn
    #   pyjwt
    #   pyopenssl
    #   service-identity
daphne==4.1.2
    # via channels
defusedxml==0.7.1
    # via python3-openid
distro==1.9.0
    # via openai
dj-pagination==2.5.0
    # via -r genflow/requirements/base.in
dj-rest-auth[with-social]==5.0.2
    # via -r genflow/requirements/base.in
django==5.1.6
    # via
    #   -r genflow/requirements/base.in
    #   channels
    #   dj-rest-auth
    #   django-allauth
    #   django-appconf
    #   django-compressor
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   drf-spectacular
django-allauth==0.57.2
    # via
    #   -r genflow/requirements/base.in
    #   dj-rest-auth
django-appconf==1.1.0
    # via django-compressor
django-compressor==4.5.1
    # via -r genflow/requirements/base.in
django-cors-headers==4.7.0
    # via -r genflow/requirements/base.in
django-filter==25.1
    # via -r genflow/requirements/base.in
djangorestframework==3.15.2
    # via
    #   -r genflow/requirements/base.in
    #   dj-rest-auth
    #   drf-spectacular
drf-spectacular==0.28.0
    # via -r genflow/requirements/base.in
h11==0.14.0
    # via httpcore
httpcore==1.0.8
    # via httpx
httpx==0.28.1
    # via openai
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.10
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   requests
    #   twisted
incremental==24.7.2
    # via twisted
inflection==0.5.1
    # via drf-spectacular
jiter==0.9.0
    # via openai
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
numpy==2.2.4
    # via pandas
oauthlib==3.2.2
    # via requests-oauthlib
openai==1.66.3
    # via -r genflow/requirements/base.in
pandas==2.2.3
    # via -r genflow/requirements/base.in
pillow==11.0.0
    # via -r genflow/requirements/base.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.22
    # via cffi
pydantic==2.11.3
    # via openai
pydantic-core==2.33.1
    # via pydantic
pyjwt[crypto]==2.10.1
    # via django-allauth
pyopenssl==25.0.0
    # via twisted
python-dateutil==2.9.0.post0
    # via pandas
python3-openid==3.2.0
    # via django-allauth
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via drf-spectacular
rcssmin==1.1.2
    # via django-compressor
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   django-allauth
    #   requests-oauthlib
    #   tiktoken
requests-oauthlib==2.0.0
    # via django-allauth
rjsmin==1.2.2
    # via django-compressor
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
service-identity==24.2.0
    # via twisted
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlparse==0.5.3
    # via django
tiktoken==0.9.0
    # via -r genflow/requirements/base.in
tqdm==4.67.1
    # via openai
twisted[tls]==24.11.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.13.2
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   pyopenssl
    #   referencing
    #   twisted
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.4.0
    # via requests
zope-interface==7.2
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
setuptools==78.1.0
    # via
    #   autobahn
    #   incremental
    #   zope-interface
